ViewThemeUtils: split talk-specific and appcompat utils
