+++ app/src/main/java/com/nextcloud/talk/adapters/items/ParticipantItem.java
@@ -155,7 +155,7 @@ public class ParticipantItem extends AbstractFlexibleItem<ParticipantItem.Partic
             if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
                 holder.binding.avatarDraweeView.getHierarchy().setPlaceholderImage(
                     DisplayUtils.getRoundedDrawable(
-                        viewThemeUtils.themePlaceholderAvatar(holder.binding.avatarDraweeView,
+                        viewThemeUtils.talk.themePlaceholderAvatar(holder.binding.avatarDraweeView,
                                                               R.drawable.ic_avatar_group)));
             } else {
                 holder.binding.avatarDraweeView.setImageResource(R.drawable.ic_circular_group);
@@ -164,7 +164,7 @@ public class ParticipantItem extends AbstractFlexibleItem<ParticipantItem.Partic
             if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
                 holder.binding.avatarDraweeView.getHierarchy().setPlaceholderImage(
                     DisplayUtils.getRoundedDrawable(
-                        viewThemeUtils.themePlaceholderAvatar(holder.binding.avatarDraweeView,
+                        viewThemeUtils.talk.themePlaceholderAvatar(holder.binding.avatarDraweeView,
                                                               R.drawable.ic_avatar_mail)));
             } else {
                 holder.binding.avatarDraweeView.setImageResource(R.drawable.ic_circular_mail);
