+++ app/src/main/java/com/nextcloud/talk/remotefilebrowser/adapters/RemoteFileBrowserItemsListViewHolder.kt
@@ -98,7 +98,7 @@ class RemoteFileBrowserItemsListViewHolder(
         binding.fileIcon
             .hierarchy
             .setPlaceholderImage(
-                viewThemeUtils.getPlaceholderImage(binding.root.context, item.mimeType)
+                viewThemeUtils.talk.getPlaceholderImage(binding.root.context, item.mimeType)
             )
 
         if (item.hasPreview) {
