+++ app/src/main/java/com/nextcloud/talk/ui/theme/viewthemeutils/AppCompatViewThemeUtils.kt
@@ -1,0 +1,93 @@
+/*
+ * Nextcloud Talk application
+ *
+ * @author Álvaro Brey
+ * Copyright (C) 2022 Álvaro Brey
+ * Copyright (C) 2022 Nextcloud GmbH
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package com.nextcloud.talk.ui.theme.viewthemeutils
+
+import android.content.res.ColorStateList
+import android.graphics.Color
+import android.widget.LinearLayout
+import androidx.appcompat.widget.SearchView
+import androidx.appcompat.widget.SwitchCompat
+import androidx.core.content.res.ResourcesCompat
+import com.nextcloud.android.common.ui.color.ColorUtil
+import com.nextcloud.android.common.ui.theme.MaterialSchemes
+import com.nextcloud.android.common.ui.theme.ViewThemeUtilsBase
+import com.nextcloud.talk.R
+import javax.inject.Inject
+
+/**
+ * View theme utils for AppCompat views (androidx.appcompat.*)
+ */
+class AppCompatViewThemeUtils @Inject constructor(schemes: MaterialSchemes, private val colorUtil: ColorUtil) :
+    ViewThemeUtilsBase(schemes) {
+    fun themeSearchView(searchView: SearchView) {
+        withScheme(searchView) { scheme ->
+            // hacky as no default way is provided
+            val editText = searchView.findViewById<SearchView.SearchAutoComplete>(R.id.search_src_text)
+            val searchPlate = searchView.findViewById<LinearLayout>(R.id.search_plate)
+            editText.textSize = SEARCH_TEXT_SIZE
+            editText.setHintTextColor(scheme.onSurfaceVariant)
+            editText.setTextColor(scheme.onSurface)
+            editText.setBackgroundColor(scheme.surface)
+            searchPlate.setBackgroundColor(scheme.surface)
+        }
+    }
+
+    fun colorSwitchCompat(switchCompat: SwitchCompat) {
+        withScheme(switchCompat) { scheme ->
+
+            val context = switchCompat.context
+
+            val thumbUncheckedColor = ResourcesCompat.getColor(
+                context.resources,
+                R.color.switch_thumb_color_unchecked,
+                context.theme
+            )
+            val trackUncheckedColor = ResourcesCompat.getColor(
+                context.resources,
+                R.color.switch_track_color_unchecked,
+                context.theme
+            )
+
+            val trackColor = Color.argb(
+                SWITCH_COMPAT_TRACK_ALPHA,
+                Color.red(scheme.primary),
+                Color.green(scheme.primary),
+                Color.blue(scheme.primary)
+            )
+
+            switchCompat.thumbTintList = ColorStateList(
+                arrayOf(intArrayOf(android.R.attr.state_checked), intArrayOf()),
+                intArrayOf(scheme.primary, thumbUncheckedColor)
+            )
+
+            switchCompat.trackTintList = ColorStateList(
+                arrayOf(intArrayOf(android.R.attr.state_checked), intArrayOf()),
+                intArrayOf(trackColor, trackUncheckedColor)
+            )
+        }
+    }
+
+    companion object {
+        private const val SWITCH_COMPAT_TRACK_ALPHA: Int = 77
+        private const val SEARCH_TEXT_SIZE: Float = 16f
+    }
+}
